Refactored MagicText to become the new SuperText
