+++ super_text/lib/super_text.dart
@@ -1,9 +1,6 @@
 export 'src/caret_layer.dart';
-export 'src/magic_text.dart';
 export 'src/super_text.dart';
 export 'src/super_text_with_selection.dart';
 export 'src/text_layout.dart';
 export 'src/text_selection_layer.dart';
 export 'src/infrastructure/blink_controller.dart';
-
-export 'src/super_duper_text.dart';
