+++ super_text/test/super_text/super_text_test_tools.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:super_text/src/magic_text.dart';
 import 'package:super_text/super_text.dart';
 
 // Everything in this file is duplicated between /test and /test_goldens because
